+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
+
+## [0.1.7] - 2019-03-21
 ### Added
 - `get_servers` method
 - `module_create` method
@@ -59,6 +61,7 @@ and this project adheres to [Semantic Versioning].
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 [Unreleased]: https://github.com/Smoose-bv/whmcspy
+[0.1.7]: https://github.com/Smoose-bv/whmcspy/releases/tag/0.1.7
 [0.1.6]: https://github.com/Smoose-bv/whmcspy/releases/tag/0.1.6
 [0.1.5]: https://github.com/Smoose-bv/whmcspy/releases/tag/0.1.5
 [0.1.4]: https://github.com/Smoose-bv/whmcspy/releases/tag/0.1.4